Separate classes into separate .cs files for clarity.
