+++ Source/HelixToolkit.Wpf.SharpDX/Controls/Effects.cs
@@ -18,207 +18,7 @@ namespace HelixToolkit.Wpf.SharpDX
     using global::SharpDX.Direct3D11;
     using global::SharpDX.D3DCompiler;
     using global::SharpDX.Direct3D;
-    using System.Runtime.InteropServices;
     using System.IO;
-
-    public sealed class RenderTechnique : IComparable
-    {
-        public RenderTechnique(string name)
-        {
-            this.Name = name;
-        }
-
-        public RenderTechnique(string name, Effect effect, InputLayout layout)
-        {
-            this.Name = name;
-            this.Device = effect.Device;
-            this.Effect = effect;
-            this.EffectTechnique = effect.GetTechniqueByName(this.Name);
-            this.InputLayout = layout;
-        }
-
-        public override string ToString()
-        {
-            return this.Name;
-        }
-
-        public override int GetHashCode()
-        {
-            return this.Name.GetHashCode();
-        }
-
-        public override bool Equals(object obj)
-        {
-            return this.Name.Equals(obj.ToString());
-        }
-
-        public int CompareTo(object obj)
-        {
-            return Name.CompareTo(obj.ToString());
-        }
-
-        public static bool operator ==(RenderTechnique a, RenderTechnique b)
-        {
-            // If both are null, or both are same instance, return true.
-            if (System.Object.ReferenceEquals(a, b))
-            {
-                return true;
-            }
-
-            // If one is null, but not both, return false.
-            if (((object)a == null) || ((object)b == null))
-            {
-                return false;
-            }
-
-            // Return true if the fields match:
-            return a.Name.Equals(b.Name);
-        }
-
-        public static bool operator !=(RenderTechnique a, RenderTechnique b)
-        {
-            return !(a == b);
-        }
-
-        public string Name { get; private set; }
-
-        public Effect Effect { get; private set; }
-
-        public EffectTechnique EffectTechnique { get; private set; }
-
-        public global::SharpDX.Direct3D11.Device Device { get; private set; }
-
-        public InputLayout InputLayout { get; private set; }
-    }
-
-    public sealed class Techniques
-    {
-        static Techniques()
-        {
-            /// <summary>
-            /// Names of techniques which are implemented by default
-            /// </summary>
-            RenderBlinn = new RenderTechnique("RenderBlinn");
-            RenderPhong = new RenderTechnique("RenderPhong");
-
-            RenderDiffuse = new RenderTechnique("RenderDiffuse");
-            RenderColors = new RenderTechnique("RenderColors");
-            RenderPositions = new RenderTechnique("RenderPositions");
-            RenderNormals = new RenderTechnique("RenderNormals");
-            RenderPerturbedNormals = new RenderTechnique("RenderPerturbedNormals");
-            RenderTangents = new RenderTechnique("RenderTangents");
-            RenderTexCoords = new RenderTechnique("RenderTexCoords");
-            RenderWires = new RenderTechnique("RenderWires");
-
-#if DEFERRED 
-            RenderDeferred = new RenderTechnique("RenderDeferred");
-            RenderGBuffer = new RenderTechnique("RenderGBuffer");
-            RenderDeferredLighting = new RenderTechnique("RenderDeferredLighting");
-            RenderScreenSpace = new RenderTechnique("RenderScreenSpace");
-#endif
-
-#if TESSELLATION 
-            RenderPNTriangs = new RenderTechnique("RenderPNTriangs");
-            RenderPNQuads = new RenderTechnique("RenderPNQuads");
-#endif
-            RenderCubeMap = new RenderTechnique("RenderCubeMap");
-            RenderLines = new RenderTechnique("RenderLines");
-            RenderPoints = new RenderTechnique("RenderPoints");
-            RenderBillboard = new RenderTechnique("RenderBillboard");
-
-            RenderTechniques = new List<RenderTechnique>
-            { 
-                RenderBlinn,
-                RenderPhong, 
-
-                RenderColors,
-                RenderDiffuse,
-                RenderPositions,
-                RenderNormals,
-                RenderPerturbedNormals,
-                RenderTangents, 
-                RenderTexCoords,
-                RenderWires,
-#if DEFERRED
-                RenderDeferred,
-                RenderGBuffer,  
-#endif
-                
-#if TESSELLATION 
-                RenderPNTriangs,
-                RenderPNQuads,
-#endif
-            };
-
-            TechniquesSourceDict = new Dictionary<RenderTechnique, byte[]>()
-            {
-                {     Techniques.RenderPhong,      Properties.Resources._default}, 
-                {     Techniques.RenderBlinn,      Properties.Resources._default}, 
-                {     Techniques.RenderCubeMap,    Properties.Resources._default}, 
-                {     Techniques.RenderColors,     Properties.Resources._default}, 
-                {     Techniques.RenderDiffuse,    Properties.Resources._default}, 
-                {     Techniques.RenderPositions,  Properties.Resources._default}, 
-                {     Techniques.RenderNormals,    Properties.Resources._default},
-                {     Techniques.RenderPerturbedNormals,    Properties.Resources._default}, 
-                {     Techniques.RenderTangents,   Properties.Resources._default}, 
-                {     Techniques.RenderTexCoords,  Properties.Resources._default}, 
-                {     Techniques.RenderWires,      Properties.Resources._default}, 
-                {     Techniques.RenderLines,      Properties.Resources._default}, 
-                {     Techniques.RenderPoints,     Properties.Resources._default},
-                {     Techniques.RenderBillboard,  Properties.Resources._default},
-    #if TESSELLATION                                        
-                {     Techniques.RenderPNTriangs,  Properties.Resources._default}, 
-                {     Techniques.RenderPNQuads,    Properties.Resources._default}, 
-    #endif 
-    #if DEFERRED            
-                {     Techniques.RenderDeferred,   Properties.Resources._deferred},
-                {     Techniques.RenderGBuffer,    Properties.Resources._deferred},
-                {     Techniques.RenderDeferredLighting , Properties.Resources._deferred},
-    #endif
-            };
-        }
-
-        internal static readonly Techniques Instance = new Techniques();
-
-        internal static readonly Dictionary<RenderTechnique, byte[]> TechniquesSourceDict;
-
-        private Techniques()
-        {
-
-        }
-
-        /// <summary>
-        /// Names of techniques which are implemented by default
-        /// </summary>
-        public static RenderTechnique RenderBlinn { get; private set; }// = new RenderTechnique("RenderBlinn");
-        public static RenderTechnique RenderPhong { get; private set; }
-
-        public static RenderTechnique RenderDiffuse { get; private set; }
-        public static RenderTechnique RenderColors { get; private set; }
-        public static RenderTechnique RenderPositions { get; private set; }
-        public static RenderTechnique RenderNormals { get; private set; }
-        public static RenderTechnique RenderPerturbedNormals { get; private set; }
-        public static RenderTechnique RenderTangents { get; private set; }
-        public static RenderTechnique RenderTexCoords { get; private set; }
-        public static RenderTechnique RenderWires { get; private set; }
-        public static RenderTechnique RenderCubeMap { get; private set; }
-        public static RenderTechnique RenderLines { get; private set; }
-        public static RenderTechnique RenderPoints { get; private set; }
-        public static RenderTechnique RenderBillboard { get; private set; }
-
-#if TESSELLATION
-        public static RenderTechnique RenderPNTriangs { get; private set; }
-        public static RenderTechnique RenderPNQuads { get; private set; }
-#endif
-
-#if DEFERRED                 
-        public static RenderTechnique RenderDeferred { get; private set; }
-        public static RenderTechnique RenderGBuffer { get; private set; }
-        public static RenderTechnique RenderDeferredLighting { get; private set; }
-        public static RenderTechnique RenderScreenSpace { get; private set; }
-#endif
-        public static IEnumerable<RenderTechnique> RenderTechniques { get; private set; }
-    }
 
     public sealed class EffectsManager : IDisposable
     {
@@ -710,43 +510,5 @@ namespace HelixToolkit.Wpf.SharpDX
 
             private Stream stream;
         }
-    }
-
-    [StructLayout(LayoutKind.Sequential, Pack = 4)]
-    public struct DefaultVertex
-    {
-        public Vector4 Position;
-        public Color4 Color;
-        public Vector2 TexCoord;
-        public Vector3 Normal;
-        public Vector3 Tangent;
-        public Vector3 BiTangent;
-
-        public const int SizeInBytes = 4 * (4 + 4 + 2 + 3 + 3 + 3);
-    }
-
-    [StructLayout(LayoutKind.Sequential, Pack = 4)]
-    public struct LinesVertex
-    {
-        public Vector4 Position;
-        public Color4 Color;
-        public const int SizeInBytes = 4 * (4 + 4);
-    }
-
-    [StructLayout(LayoutKind.Sequential, Pack = 4)]
-    public struct CubeVertex
-    {
-        public Vector4 Position;
-        public const int SizeInBytes = 4 * 4;
-    }
-
-    [StructLayout(LayoutKind.Sequential, Pack = 4)]
-    public struct BillboardVertex
-    {
-        public Vector4 Position;
-        public Color4 Color;
-        public Vector4 TexCoord;
-        //public Vector2 Offset;
-        public const int SizeInBytes = 4 * (4 + 4 + 4);
     }
 }
