+++ Source/HelixToolkit.Wpf.SharpDX/HelixToolkit.Wpf.SharpDX.csproj
@@ -145,6 +145,9 @@
     <Compile Include="Controls\DPFCanvas.cs" />
     <Compile Include="Controls\IRenderable.cs" />
     <Compile Include="Controls\IRenderHost.cs" />
+    <Compile Include="Controls\Layouts.cs" />
+    <Compile Include="Controls\RenderTechnique.cs" />
+    <Compile Include="Controls\Techniques.cs" />
     <Compile Include="Controls\Trackball.cs" />
     <Compile Include="Controls\Viewport3DX.cs" />
     <Compile Include="Core\Box3D.cs" />
